Reorganize file structure
This commit reorganizes the `test/` and `src/` directory since they were getting a bit crowded. Both are sorted by category with the following new folders:

- google-analytics
- process-results
- publish
