+++ src/google-analytics/client.js
