+++ src/google-analytics/credential-loader.js
@@ -1,4 +1,4 @@
-const config = require("./config")
+const config = require("../config")
 
 global.analyticsCredentialsIndex = 0
