+++ src/google-analytics/query-builder.js
@@ -1,4 +1,4 @@
-const config = require('./config')
+const config = require('../config')
 
 const buildQuery = (report) => {
   let query = Object.assign({}, report.query)
