+++ src/process-results/result-totals-calculator.js
