+++ test/support/mocks/googleapis-analytics.js
