+++ test/support/mocks/googleapis-auth.js
